Now you can see your own feeds on frontpage. Added shareable url highlight to feeds description
Fixed errorhandler for forms (multiple errors etc.)
Prefill users form inputs on error.
Due to herokus ephemeral filesystem, deletin items only if they exist

+++ feeds.py
@@ -24,7 +24,8 @@ def delete_feed(secret):
     items = db.session.execute(sql, {"feed_id":feed['id'] }).fetchall()
 
     for item in items:
-        os.remove(item['path'])
+        if os.path.exists(item['path']):
+            os.remove(item['path'])
 
     sql = "DELETE from items WHERE feed_id=:feed_id"
     db.session.execute(sql, {"feed_id":feed['id'] })
@@ -38,6 +39,11 @@ def get_feed(secret):
     sql = "SELECT * FROM feeds WHERE secret=:secret"
     result = db.session.execute(sql, {"secret":secret})
     return result.fetchone()
+
+def get_feeds():
+    sql = "SELECT * FROM feeds WHERE user_id=:user_id"
+    result = db.session.execute(sql, { "user_id":users.user_id() })
+    return result.fetchall()
 
 def get_items(secret):
     sql = "SELECT I.id as id, I.description as description, I.path as path, I.created_at as created_at, I.user_id as user_id FROM items I, feeds F WHERE I.feed_id=F.id AND F.secret=:secret ORDER BY I.created_at DESC"
@@ -72,13 +78,17 @@ def delete_item(secret, id):
     item = result.fetchone()
     if feed['user_id'] != user_id and item['user_id'] != user_id:
         return False
-    os.remove(item['path'])
+    if os.path.exists(item['path']):
+        os.remove(item['path'])
     sql = "DELETE FROM items WHERE feed_id=:feed_id AND id=:id"
     db.session.execute(sql, {"feed_id":feed[0], "id":id })
     db.session.commit()
     return True
-
-
+    
+def get_votes(item_id):
+    sql = "SELECT id FROM votes WHERE feed_id=:feed_id AND item_id=:item_id"
+    result = db.session.execute(sql, {"secret":secret})
+    return result.fetchall()
 
 def vote_item(secret, item_id):
     user_id = users.user_id()
